criteriaThresholds v1.2: eliminating empty XMCDA tags in the output
If a value for (criterion, threshold) is not specified (left empty) in
the csv file, the XMCDA tag should not be produced in the output.
Same for the 'scale' sub-tag of <criterion>, which is present only if
the corresponding value in the csv is not empty.

Added tests, updated documentation.